nice regular status updates at the end of AISpecs
the trick is to apply the overall time limit to the ROSpec, then make the AISpec
a short duration-controlled repetitive task that will be executed repeatedly
within the ROSpec execution.  see section 6 of the LLRP 1.0.1 spec.

this kind of invalidates the comment of 861cf6af517adefbcb1e306476e47ed5d96c1dd4