Better project setup.
We want to have the license be called license.
Switched from old distutils to much better setuptools.
Made 'inventory' public in module so that we can have it as a script on the command line (look console_scripts in setup.py).
Much more comprehensive .gitignore to ignore build packages and distributions etc.