remove ability to specify mode-index at all
it's so confusing!  the LLRP spec says ModeIndex is "an index into the
UHFC1G2RFModeTable," but the only values readers accept for the ModeIndex
parameters -- at least, Impinj Speedway and Motorola MC9190 -- in fact refer to
ModeIdentifier values.

on an Impinj reader, supply ModeIndex=1000 and see that this conflation is all
too real!

pertinent to #63