Fixes decode_FrequencyInformation that was broken, especially when hopping was false.
There was a typo in the decode function to use for decoding FixedFrequencyTable.

There was also a bug with unpack in FixedFrequencyTable and FrequencyHopTable
as unpack return a tuple even for a single element.

Example of good wrong output for an ETSI device:
u'UHFBandCapabilities': {u'FrequencyInformation': {u'Hopping': False},

Example of good output for the same device:
u'UHFBandCapabilities': {u'FrequencyInformation': {u'FixedFrequencyTable': {u'Frequency1': 865700,
                                                                            u'Frequency2': 866300,
                                                                            u'Frequency3': 866900,
                                                                            u'Frequency4': 867500,
                                                                            u'NumFrequencies': 4},
                                                   u'Hopping': False},