Add a debugfast function to increase parsing speed
Even if log level is higher than "DEBUG", using debug() still incurs a
performance cost just to determine that it can be ignored.
And we have a lot of logging on the parsing path of a tagreport event.

The purpose of this change is to add and use a new "debugfast()"
function that will more efficiently skip calling "logging.debug()" if
not needed.

In my empiric test, this change can almost reduce by 2 the time to parse
a single TagReport.