Fixes fviard/sllurp#1 - set reader state back to disconnected in case of lost connection
If the connection is lost, set the reader llrp state back to
DISCONNECTED before attempting a reconnection.
And so, if reconnection is successful, the llrp configuration of the
reader will be reset. ie the config + rospec will be pushed again to the
reader.
This solves the case of readers that are "rebooted" with
auto-reconnection enabled.