llrp.py: Raise errors that are encountered during msg deserialize or handle_message and ensure ReaderConfigurationErrors are fatals
Before, when the reader was replying with an error to an early
operation, like "ADD_ROSPEC" to start an inventory, the error was
ignored, and so the situation was frozen:
the connection was kept and the reader was idle as command was failed,
but from the client point of view, we were already in an inventory stage
and just waiting for tag reads.

Now, if such a reader configuration error is encountered, the connection
will be terminated and so the client can notice that the reader is not
alive anymore.

Unresolved TODO: We should still find a way to notify the client about
the ReaderConfigurationError.